Add upstream documentation links to Database.SQLite3 like we have in .Direct
Also, move documentation of 'exec' and 'prepare' (other than the link)
from .Direct to Database.SQLite3.  For each function, we want documentation
that applies to every level of abstraction to be in Database.SQLite3, where
more users can see it.  For .Direct and .Bindings, we only want information
that pertains to that level.