Write custom Show instances for ParamIndex and ColumnIndex so they don't show the data constructor
ParamIndex and ColumnIndex are supposed to look and feel like regular integers.
With the default Show instances, this would have a hidden quirk:

    stmt <- prepare conn "SELECT ?, ?, ?"
    count <- bindParameterCount stmt
    putStrLn $ "This query has " ++ show count ++ " parameter(s)"

This would print "This query has ParamIndex 3 parameter(s)".

Here are a few other newtypes that might benefit from a Show instance
like this, but are less user-facing:

 * Utf8

 * CError

 * CColumnType

 * CNumBytes