Add sqlite3_sql & statementText
Adding accessors for getting the SQL statement text from an sqlite
statement.

This can be used for debugging purposes -- say formatting error
messages with SQL in them without having to pass around the original
query string used to build up the statement.

Decided to call the Direct entry point 'statementText' instead of just
'sql' because the latter is probably a common name in user-code and
we'd get compiler warnings on binding shadowing.