Add sqlite3.h and sqlite3ext.h to install-includes
sqlite3ext.h is the header for sqlite3 extensions to include. It was obtained from
https://raw.githubusercontent.com/mackyle/sqlite/version-3.15.2/src/sqlite3ext.h

This patch allows the aforementioned headers to be found by ghc by adding a cabal
dependency to direct-sqlite. I have found this useful when compiling extensions
written in c against the sqlite version that direct-sqlite bundles.